% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_operator.R
\name{\%=~\%}
\alias{\%=~\%}
\title{Regex Match Operator}
\usage{
lhs \%=~\% rhs
}
\arguments{
\item{lhs}{left-hand side argument}

\item{rhs}{right-hand side argument}
}
\value{
Logical
}
\description{
Regex Match Operator
}
